move package_generator to rdm/gen/package class
problem: 
- we want to introduce more elaborate generators soon

solution: 
- give them their own namespace, that maps directly to CLI classes
